update the User model to accept requesting_actor_id when saving
Previously, the requesting entity's authz side id was set on the object
prior to save or update. This mixed concerns, was complicated, and may
have been implicated in some bugs observed in prod.

Now, the authz id of the requestor will be passed as an argument to
User#save_as and #save will be private. This is much simpler and doesn't
pollute the data model with irrelevant concerns.