Extract org authz policy via a DSL
OrgAuthPolicy provides a DSL to declare the default objects and their
ACEs for an organization. Also reduces the number of places that depend
on a given DAL (e.g., CouchRest vs. SQL/mappers). As a bonus, local
caching is used to cut down the number of database requests required to
set up an org--unassigned orgs can now be created in ~20-25s.