Use `login_attribute` for LDAP record mapping
Since `objectsid` is not part of the standard LDAP record schema we
cannot count on it's contents being present (or valid) as a UUID for
mapping Chef records to LDAP records. Same goes for `uid` in POSIX
LDAP schemas like OpenLDAP. The safest bet is to just use the value
as determined to the configurable `login_attribute` which defaults
to 'samaccountname'.