Fix bug in handling non-org-scoped requests
The bug manifests for users/:id (show) where :id will be the username
and orgname will return username. Now that we check for clients first,
this would cause an error trying to talk to couch where no such db
exists. With this patch, we'll get an ArgumentError which is rescued
and we move on to checking for a user.