Lock down pg gem, remove mysql
We no longer support MySQL, so we can remove it.  Since we no longer
need to segregate database-related gems (since we now only have one),
we can remove the 'pg' gem group, as well.

This is the latest version of the pg gem, which has better support for
compiling the native extensions against non-standard library and
header locations, which is useful in Omnibus package builds.