Thread org_name through mappers.
This bundles up some cleanups of the usage of the mappers, along with
threading the org name through. This is needed since we don't have an
easy org_id to org name mapping in SQL, and the API expects the
group record to to have the org name in it.