Allow mapper to override default value of hash_type
Since we are setting the default value of hash_type to 'bcrypt', and
we need to load it from the data store correctly, we put in this special
case to explicitly override hash_type using the db data.

For some reason, the loading code will not overwrite @hash_type if there
is something already in there. It may be a bug. The real solution is to
add in semantics to support default values, however, considering that
this code is not living for much longer, this may not be worth pursing.

Noting here, this solution is a kludge.