Compat fixes for dealing with public keys and certificates for users
When reading from the db, we have a single column that will contain
either a certificate or a public key and the mapping code will give
the result to us with the data mapped to a "certificate" in the data
hash.

We now allow public_key and certificate to be set and detect the key
type using a simple rule based on the BEGIN header present in the PEM
format.